ci(build): build check config have been updated
This commit fixes a bug in the build script that was causing errors when trying to deploy to GitHub Pages. The bug was related to a leftover code that attempted to create a build for GitHub Pages, even though this feature has been disabled in previous builds. The code has been removed and the build script has been simplified.